improvements

+++ camera.h
@@ -1,11 +1,11 @@
 #ifndef CAMERA_H
 #define CAMERA_H
 
-#include "vec3.h"
 #include "hittable.h"
+#include "vec3.h"
 
 typedef struct camera {
-	double aspect_ratio;
+	float aspect_ratio;
 	int image_width;
 	int image_height;
 	vec3_t center;
@@ -14,12 +14,12 @@ typedef struct camera {
 	vec3_t pixel_dv;
 	int samples_per_pixel;
 	int max_depth;
-	double vfov;
+	float vfov;
 	vec3_t lookfrom;
 	vec3_t lookat;
 	vec3_t vup;
-	double defocus_angle;
-	double focus_dist;
+	float defocus_angle;
+	float focus_dist;
 	vec3_t defocus_disk_u;
 	vec3_t defocus_disk_v;
 } camera_t;
