+++ interval.h
@@ -2,23 +2,23 @@
 #define INTERVAL_H
 
 typedef struct interval {
-	double min;
-	double max;
+	float min;
+	float max;
 } interval_t;
 
-static inline double interval_size(interval_t i) { return i.max - i.min; }
+static inline float interval_size(interval_t i) { return i.max - i.min; }
 
-static inline int interval_contains(interval_t i, double x)
+static inline int interval_contains(interval_t i, float x)
 {
 	return i.min <= x && x <= i.max;
 }
 
-static inline int interval_surrounds(interval_t i, double x)
+static inline int interval_surrounds(interval_t i, float x)
 {
 	return i.min < x && x < i.max;
 }
 
-static inline double interval_clamp(interval_t i, double x)
+static inline float interval_clamp(interval_t i, float x)
 {
 	if (x < i.min)
 		return i.min;
