+++ vec3.h
@@ -5,32 +5,32 @@
 #include <stdio.h>
 
 typedef struct vec3 {
-	double x, y, z;
+	float x, y, z;
 } vec3_t;
 
-vec3_t vec3_random(double min, double max);
+vec3_t vec3_random(float min, float max);
 vec3_t vec3_random_unit_vector();
 vec3_t vec3_random_on_hemisphere(vec3_t normal);
 vec3_t vec3_random_in_unit_disk();
 
-static inline double vec3_length_squared(vec3_t u)
+static inline float vec3_length_squared(vec3_t u)
 {
 	return u.x * u.x + u.y * u.y + u.z * u.z;
 }
 
-static inline double vec3_length(vec3_t u)
+static inline float vec3_length(vec3_t u)
 {
-	return sqrt(vec3_length_squared(u));
+	return sqrtf(vec3_length_squared(u));
 }
 
-static inline vec3_t vec3_scaled(vec3_t u, double t)
+static inline vec3_t vec3_scaled(vec3_t u, float t)
 {
 	return (vec3_t){u.x * t, u.y * t, u.z * t};
 }
 
 static inline vec3_t vec3_normalized(vec3_t u)
 {
-	double l;
+	float l;
 
 	l = vec3_length(u);
 
@@ -38,7 +38,7 @@ static inline vec3_t vec3_normalized(vec3_t u)
 }
 
 vec3_t vec3_reflected(vec3_t u, vec3_t n);
-vec3_t vec3_refracted(vec3_t u, vec3_t n, double etai_over_etat);
+vec3_t vec3_refracted(vec3_t u, vec3_t n, float etai_over_etat);
 int vec3_near_zero(vec3_t u);
 
 static inline vec3_t vec3_add(vec3_t u, vec3_t v)
@@ -51,7 +51,7 @@ static inline vec3_t vec3_sub(vec3_t u, vec3_t v)
 	return (vec3_t){u.x - v.x, u.y - v.y, u.z - v.z};
 }
 
-static inline double vec3_dot(vec3_t u, vec3_t v)
+static inline float vec3_dot(vec3_t u, vec3_t v)
 {
 	return u.x * v.x + u.y * v.y + u.z * v.z;
 }
